w2/requirements.txt
django
djangorestframework
django_filter
psycopg2-binary